requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
pillow
pytest
pytest-asyncio
httpx
mock